requirements.txt
click==7.1.2
Flask==1.1.2
img2pdf==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
pikepdf==5.1.2
Pillow==8.1.2
PyJWT==2.0.1
python-dotenv==0.15.0
Werkzeug==1.0.1
